pyproject.toml
[project]
name = "my-selenium-tests"
version = "0.1.0"
description = "selenium test automation"
authors = [
    {name = "akbarm",email = "inzamam.akbar@netsoltech.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest (>=8.4.1,<9.0.0)",
    "mailslurp-client (>=16.2.1,<17.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "allure-pytest (>=2.14.3,<3.0.0)"
]
[tool.pytest.ini_options]
addopts = "--alluredir=allure-results"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
